requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
scikit-learn
numpy
pandas
python-dotenv
joblib
scipy
